+++ NEWS.md
@@ -1,5 +1,10 @@
 # News for Package `cardelino`
 
+## Changes in v0.6.3 (18/08/2019)
+* Add devarianceIC() for Devariance information criterion (DIC) for 
+  clone_id_Gibbs()
+* Change the default of relax_Config=TRUE for clone_id() and clone_id_Gibbs()
+
 ## Changes in version 0.6.2
 * Fix the bug in `Geweke_Z` function for convegence diagnostic
 * Change the default number of iterations in `clone_id` to 5000